+++ src/main/java/com/otus/bookstore/service/impl/CommentServiceImpl.java
@@ -1,23 +1,25 @@
 package com.otus.bookstore.service.impl;
 
-import com.otus.bookstore.exception.CommentErrorSavedException;
-import com.otus.bookstore.exception.CommentNotFoundException;
+import com.otus.bookstore.exception.EntitySaveException;
 import com.otus.bookstore.model.BookComment;
 import com.otus.bookstore.model.BookCommentId;
 import com.otus.bookstore.model.Comment;
 import com.otus.bookstore.repository.BookCommentRepository;
 import com.otus.bookstore.repository.CommentRepository;
 import com.otus.bookstore.service.CommentService;
-import jakarta.transaction.Transactional;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import java.security.InvalidParameterException;
 import java.util.List;
 import java.util.Optional;
 
-@Transactional
 @Service
 public class CommentServiceImpl implements CommentService {
+    public static final String ERROR_NOT_FOUND_AUTHOR = "Comment with id %s not found";
+
     private final CommentRepository commentRepository;
     private final BookCommentRepository bookCommentRepository;
 
@@ -29,7 +31,8 @@ public class CommentServiceImpl implements CommentService {
     }
 
     @Override
-    public Comment save(Comment comment) throws CommentErrorSavedException {
+    @Transactional
+    public Comment save(Comment comment) {
         try {
             comment = commentRepository.save(comment);
 
@@ -45,49 +48,51 @@ public class CommentServiceImpl implements CommentService {
 
             bookCommentRepository.save(bookComment);
             return comment;
-        } catch (Exception e) {
-            throw new CommentErrorSavedException();
+        } catch (RuntimeException e) {
+            throw new EntitySaveException(e.getMessage(), e);
         }
     }
 
     @Override
+    @Transactional(readOnly = true)
     public List<Comment> getAll() {
         return commentRepository.findAll();
     }
 
     @Override
-    public Optional<Comment> findById(Long id) throws CommentNotFoundException {
-        try {
-            Optional<Comment> commentOptional = commentRepository.findById(id);
-
-            if (commentOptional.isEmpty()) {
-                throw new CommentNotFoundException(String.format(CommentNotFoundException.ERROR_NOT_FOUND, id));
-            }
-
-            return commentOptional;
-        } catch (RuntimeException e) {
-            System.out.println("Comment not found 2");
-            throw new CommentNotFoundException(String.format(CommentNotFoundException.ERROR_NOT_FOUND, id));
+    @Transactional(readOnly = true)
+    public Optional<Comment> findById(Long id) {
+        if (id <= 0) {
+            throw new InvalidParameterException(String.format(ERROR_NOT_FOUND_AUTHOR, id));
         }
+
+        Optional<Comment> commentOptional = commentRepository.findById(id);
+
+        if (commentOptional.isEmpty()) {
+            throw new EntityNotFoundException(String.format(ERROR_NOT_FOUND_AUTHOR, id));
+        }
+
+        return commentOptional;
     }
 
     @Override
-    public void deleteById(Long id) throws CommentNotFoundException {
-        try {
-            Optional<Comment> findComment = commentRepository.findById(id);
-
-            if (findComment.isEmpty()) {
-                throw new CommentNotFoundException(String.format(CommentNotFoundException.ERROR_NOT_FOUND, id));
-            }
-
-            commentRepository.deleteById(id);
-
-            bookCommentRepository.deleteByBookCommentId(BookCommentId.builder()
-                    .bookId(findComment.get().getBook().getId())
-                    .commentId(findComment.get().getId())
-                    .build());
-        } catch (RuntimeException e) {
-            throw new CommentNotFoundException(String.format(CommentNotFoundException.ERROR_NOT_FOUND, id));
+    @Transactional
+    public void deleteById(Long id) {
+        if (id <= 0) {
+            throw new InvalidParameterException(String.format(ERROR_NOT_FOUND_AUTHOR, id));
         }
+
+        Optional<Comment> findComment = commentRepository.findById(id);
+
+        if (findComment.isEmpty()) {
+            throw new EntityNotFoundException(String.format(ERROR_NOT_FOUND_AUTHOR, id));
+        }
+
+        commentRepository.deleteById(id);
+
+        bookCommentRepository.deleteById(BookCommentId.builder()
+                .bookId(findComment.get().getBook().getId())
+                .commentId(findComment.get().getId())
+                .build());
     }
 }
